models/minigpt4/requirements.txt
accelerate==0.21.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bitsandbytes==0.40.2
braceexpand==0.1.7
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
decorator==5.1.1
decord==0.6.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
executing==1.2.0
fairscale==0.4.13
fastapi==0.103.2
filelock==3.12.3
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.9.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.34
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.18.0
idna==3.4
imageio==2.31.6
iopath==0.1.10
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
lit==16.0.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.7
networkx==3.1
nh3==0.2.14
nm==0.0.4
numpy==1.25.2
packaging==23.1
parso==0.8.3
pathtools==0.1.2
peft==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pydantic==1.10.13
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
rich==13.6.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.30.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tiktoken==0.5.1
timm==0.6.13
tokenizers==0.13.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
wandb==0.15.10
wavedrom==2.0.3.post3
wcwidth==0.2.6
webdataset==0.2.48
yarl==1.9.2
omegaconf==2.3.0
opencv-python==4.7.0.72
matplotlib==3.7.0
matplotlib-inline==0.1.6
visual-genome==1.1.1
pandas==2.1.0
python-multipart==0.0.6